Move `check_cache` function into brian2/__init__.py and add a `clear_cache` function
